+++ Sources/MapboxDirections/VisualInstructionComponent.swift
@@ -35,7 +35,12 @@ public extension VisualInstruction {
         case image(image: ImageRepresentation, alternativeText: TextRepresentation)
         
         /**
-         The compoment contains the localized word for “Exit”.
+         The component is an image of a zoomed junction, with a fallback text representation.
+         */
+        case guidanceView(image: GuidanceViewImageRepresentation, alternativeText: TextRepresentation)
+        
+        /**
+         The component contains the localized word for “Exit”.
          
          This component may appear before or after an `.exitCode` component, depending on the language. You can hide this component if the adjacent `.exitCode` component has an obvious exit-number appearance, for example with an accompanying [motorway exit icon](https://commons.wikimedia.org/wiki/File:Sinnbild_Autobahnausfahrt.svg).
          */
@@ -149,6 +154,22 @@ public extension VisualInstruction.Component {
     }
 }
 
+/// A guidance view image representation of a visual instruction component.
+public struct GuidanceViewImageRepresentation: Equatable {
+    /**
+     Initializes an image representation bearing the image at the given URL.
+     */
+    public init(imageURL: URL?) {
+        self.imageURL = imageURL
+    }
+
+    /**
+     Returns a remote URL to the image file that represents the component.
+     */
+    public let imageURL: URL?
+    
+}
+
 extension VisualInstruction.Component: Codable {
     private enum CodingKeys: String, CodingKey {
         case kind = "type"
@@ -156,6 +177,7 @@ extension VisualInstruction.Component: Codable {
         case abbreviatedText = "abbr"
         case abbreviatedTextPriority = "abbr_priority"
         case imageBaseURL
+        case imageURL = "url"
         case directions
         case isActive = "active"
     }
@@ -164,6 +186,7 @@ extension VisualInstruction.Component: Codable {
         case delimiter
         case text
         case image = "icon"
+        case guidanceView = "guidance-view"
         case exit
         case exitCode = "exit-number"
         case lane
@@ -203,6 +226,13 @@ extension VisualInstruction.Component: Codable {
             self = .exitCode(text: textRepresentation)
         case .lane:
             preconditionFailure("Lane component should have been initialized before decoding text")
+        case .guidanceView:
+            var imageURL: URL?
+            if let imageURLString = try container.decodeIfPresent(String.self, forKey: .imageURL) {
+                imageURL = URL(string: imageURLString)
+            }
+            let guidanceViewImageRepresentation = GuidanceViewImageRepresentation(imageURL: imageURL)
+            self = .guidanceView(image: guidanceViewImageRepresentation, alternativeText: textRepresentation)
         }
     }
     
@@ -232,6 +262,10 @@ extension VisualInstruction.Component: Codable {
             textRepresentation = .init(text: "", abbreviation: nil, abbreviationPriority: nil)
             try container.encode(indications, forKey: .directions)
             try container.encode(isUsable, forKey: .isActive)
+        case .guidanceView(let image, let alternativeText):
+            try container.encode(Kind.guidanceView, forKey: .kind)
+            textRepresentation = alternativeText
+            try container.encodeIfPresent(image.imageURL?.absoluteString, forKey: .imageURL)
         }
         
         if let textRepresentation = textRepresentation {
@@ -254,6 +288,10 @@ extension VisualInstruction.Component: Equatable {
               let .image(rhsURL, rhsAlternativeText)):
             return lhsURL == rhsURL
                 && lhsAlternativeText == rhsAlternativeText
+        case (let .guidanceView(lhsURL, lhsAlternativeText),
+              let .guidanceView(rhsURL, rhsAlternativeText)):
+            return lhsURL == rhsURL
+                && lhsAlternativeText == rhsAlternativeText
         case (let .lane(lhsIndications, lhsIsUsable),
               let .lane(rhsIndications, rhsIsUsable)):
             return lhsIndications == rhsIndications
@@ -263,6 +301,7 @@ extension VisualInstruction.Component: Equatable {
              (.image, _),
              (.exit, _),
              (.exitCode, _),
+             (.guidanceView, _),
              (.lane, _):
             return false
         }